+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 * Add configuration in administration
 * Add option to revoke connections in personal user profile
 * Add custom login grant
+* Add Microsoft provider
+* Add Atlassian provider